shftrs

+++ simulator.c
@@ -1569,7 +1569,7 @@ int main (int argc, char** argv) {
             break;
         }
         case 0x7c: {//shftrs value
-            int16_t toShift = ((int16_t)memory[sp - 2] << 8 | (int16_t)memory[sp - 1]);
+            int16_t toShift = ((uint8_t)memory[sp - 2] << 8 | (uint8_t)memory[sp - 1]);
             toShift = toShift >> memory[pc + 1];
             memory[sp - 1] = toShift & 0xFF; // least significant
             memory[sp - 2] = (toShift >> 8) & 0xFF; //most significant
